% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gadget.R
\name{my_gadget}
\alias{my_gadget}
\title{An example shinygadget}
\usage{
my_gadget(min, max, theme = "flatly")
}
\arguments{
\item{min}{The minimum x value}

\item{max}{The maximum x value}

\item{theme}{A \code{character} string for a theme provided
by the \code{shinythemes} package}
}
\description{
This would normally be an exciting description
}
