% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{numberize}
\alias{numberize}
\title{Scale and One-Hot Encode Data Frames}
\usage{
numberize(df)
}
\arguments{
\item{df}{a data frame for conversion}
}
\value{
a data frame
}
\description{
Takes a data frame and converts all numeric variables to a
scale from 0 to 1 and converts all character and factor variables to a
one-hot encoded structure
}
\examples{

\dontrun{
numberize(iris)
}

}
