% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{password}
\alias{password}
\title{Generate a Random Password}
\usage{
password()
}
\value{
a string
}
\description{
samples with replacement from a group of possible uppercase
letters, lowercase letters, numbers, and special characters to output a
string that could be used as a password
}
\examples{
\dontrun{
password()
}

}
